+++ tests/Main.hs
@@ -22,6 +22,7 @@ import qualified Operations.Derive
 import qualified Operations.Filter
 import qualified Operations.GroupBy
 import qualified Operations.InsertColumn
+import qualified Operations.Join
 import qualified Operations.Merge
 import qualified Operations.ReadCsv
 import qualified Operations.Sort
@@ -4960,6 +4961,7 @@ tests =
             ++ Operations.Filter.tests
             ++ Operations.GroupBy.tests
             ++ Operations.InsertColumn.tests
+            ++ Operations.Join.tests
             ++ Operations.Merge.tests
             ++ Operations.ReadCsv.tests
             ++ Operations.Sort.tests